change tab-logic
