+++ src/components/TabList/TabList.jsx
@@ -1,10 +1,19 @@
 'use client';
 
+import { useState } from 'react';
 import PropTypes from 'prop-types';
 
 import { TabItem } from '../TabItem';
 
 export const TabList = ({ data, isBenefit = false, className }) => {
+  const [open, setOpen] = useState(null);
+
+  const toggle = index => {
+    if (open === index) {
+      return setOpen(null);
+    }
+    setOpen(index);
+  };
   return (
     <ul
       className={`${
@@ -12,7 +21,13 @@ export const TabList = ({ data, isBenefit = false, className }) => {
       } ${className}`}
     >
       {data.map((data, index) => (
-        <TabItem key={index} isBenefit={isBenefit} data={data} />
+        <TabItem
+          key={index}
+          open={index === open}
+          isBenefit={isBenefit}
+          data={data}
+          toggle={() => toggle(index)}
+        />
       ))}
     </ul>
   );
